+++ src/main/java/kr/co/morandi/backend/domain/defensemanagement/management/request/StartDailyDefenseServiceRequest.java
@@ -1,0 +1,20 @@
+package kr.co.morandi.backend.domain.defensemanagement.management.request;
+
+import lombok.AccessLevel;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+
+import java.time.LocalDateTime;
+
+@Getter
+@NoArgsConstructor(access = AccessLevel.PROTECTED)
+public class StartDailyDefenseServiceRequest {
+
+    private Long problemNumber;
+
+    @Builder
+    private StartDailyDefenseServiceRequest(LocalDateTime requestDateTime, Long problemNumber) {
+        this.problemNumber = problemNumber;
+    }
+}
