+++ src/main/java/kr/co/morandi/backend/domain/defensemanagement/session/model/ExamStatus.java
@@ -1,0 +1,5 @@
+package kr.co.morandi.backend.domain.defensemanagement.session.model;
+
+public enum ExamStatus {
+    IN_PROGRESS, COMPLETED, STAGE_PROGRESS;
+}
