--- src/main/java/kr/co/morandi/backend/domain/exammanagement/session/model/ExamStatus.java
@@ -1,5 +1,0 @@
-package kr.co.morandi.backend.domain.exammanagement.session.model;
-
-public enum ExamStatus {
-    IN_PROGRESS, COMPLETED, STAGE_PROGRESS;
-}
