+++ src/main/java/kr/co/morandi/backend/infrastructure/persistence/defensemanagement/sessiondetail/SessionDetailRepository.java
@@ -1,6 +1,6 @@
 package kr.co.morandi.backend.infrastructure.persistence.defensemanagement.sessiondetail;
 
-import kr.co.morandi.backend.domain.exammanagement.sessiondetail.model.SessionDetail;
+import kr.co.morandi.backend.domain.defensemanagement.sessiondetail.model.SessionDetail;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface SessionDetailRepository extends JpaRepository<SessionDetail, Long> {
